Backend/app/requirement.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
joblib
yfinance
ta
lightgbm
python-dotenv
nltk
